+++ upcloud/managedobjectstorage/resource_upcloud_managed_object_storage_test.go
@@ -1,21 +1,26 @@
 package managedobjectstoragetests
 
 import (
+	"fmt"
 	"regexp"
 	"strings"
 	"testing"
 
 	"github.com/UpCloudLtd/terraform-provider-upcloud/internal/utils"
 	"github.com/UpCloudLtd/terraform-provider-upcloud/upcloud"
+	"github.com/hashicorp/terraform-plugin-testing/config"
+	"github.com/hashicorp/terraform-plugin-testing/helper/acctest"
 	"github.com/hashicorp/terraform-plugin-testing/helper/resource"
 	"github.com/hashicorp/terraform-plugin-testing/plancheck"
 )
+
+const managedObjectStorageResource = "upcloud_managed_object_storage.this"
 
 func TestAccUpcloudManagedObjectStorage(t *testing.T) {
 	testDataS1 := utils.ReadTestDataFile(t, "testdata/managed_object_storage_s1.tf")
 	testDataS2 := utils.ReadTestDataFile(t, "testdata/managed_object_storage_s2.tf")
 
-	this := "upcloud_managed_object_storage.this"
+	this := managedObjectStorageResource
 	minimal := "upcloud_managed_object_storage.minimal"
 	bucket := "upcloud_managed_object_storage_bucket.this"
 
@@ -119,7 +124,7 @@ func TestAccUpcloudManagedObjectStorage_CustomDomain(t *testing.T) {
 	testDataS2 := utils.ReadTestDataFile(t, "testdata/managed_object_storage_custom_domain_s2.tf")
 	testDataS3 := utils.ReadTestDataFile(t, "testdata/managed_object_storage_custom_domain_s3.tf")
 
-	objsto := "upcloud_managed_object_storage.this"
+	objsto := managedObjectStorageResource
 	customDomain := "upcloud_managed_object_storage_custom_domain.this"
 
 	resource.ParallelTest(t, resource.TestCase{
@@ -156,6 +161,105 @@ func TestAccUpcloudManagedObjectStorage_CustomDomain(t *testing.T) {
 					PreApply: []plancheck.PlanCheck{
 						plancheck.ExpectResourceAction(objsto, plancheck.ResourceActionNoop),
 						plancheck.ExpectResourceAction(customDomain, plancheck.ResourceActionDestroy),
+					},
+				},
+			},
+		},
+	})
+}
+
+// TestAccUpcloudManagedObjectStorage_FullNetworkReplace tests that a complete network replacement
+// (no overlap between old and new set) succeeds. The two-step update sends an empty intermediate
+// network list, which verifies the API accepts it before adding the new networks.
+func TestAccUpcloudManagedObjectStorage_FullNetworkReplace(t *testing.T) {
+	testDataS1 := utils.ReadTestDataFile(t, "testdata/managed_object_storage_full_network_replace_s1.tf")
+	testDataS2 := utils.ReadTestDataFile(t, "testdata/managed_object_storage_full_network_replace_s2.tf")
+
+	objsto := managedObjectStorageResource
+	prefix := fmt.Sprintf("tf-acc-test-objsto-fullswap-%s-", acctest.RandString(4))
+	octet := acctest.RandIntRange(10, 250)
+	variables := map[string]config.Variable{
+		"prefix": config.StringVariable(prefix),
+		"cidr_a": config.StringVariable(fmt.Sprintf("172.%d.1.0/24", octet)),
+		"cidr_b": config.StringVariable(fmt.Sprintf("172.%d.2.0/24", octet)),
+	}
+
+	resource.ParallelTest(t, resource.TestCase{
+		PreCheck: func() { upcloud.TestAccPreCheck(t) },
+		Steps: []resource.TestStep{
+			{
+				ProtoV6ProviderFactories: upcloud.TestAccProviderFactories,
+				Config:                   testDataS1,
+				ConfigVariables:          variables,
+			},
+			{
+				ProtoV6ProviderFactories: upcloud.TestAccProviderFactories,
+				Config:                   testDataS2,
+				ConfigVariables:          variables,
+				ConfigPlanChecks: resource.ConfigPlanChecks{
+					PreApply: []plancheck.PlanCheck{
+						plancheck.ExpectResourceAction(objsto, plancheck.ResourceActionUpdate),
+					},
+				},
+				Check: resource.ComposeTestCheckFunc(
+					resource.TestCheckResourceAttr(objsto, "network.#", "1"),
+				),
+			},
+			{
+				ProtoV6ProviderFactories: upcloud.TestAccProviderFactories,
+				Config:                   testDataS2,
+				ConfigVariables:          variables,
+				ConfigPlanChecks: resource.ConfigPlanChecks{
+					PreApply: []plancheck.PlanCheck{
+						plancheck.ExpectResourceAction(objsto, plancheck.ResourceActionNoop),
+					},
+				},
+			},
+		},
+	})
+}
+
+func TestAccUpcloudManagedObjectStorage_NetworkChange(t *testing.T) {
+	testDataS1 := utils.ReadTestDataFile(t, "testdata/managed_object_storage_network_change_s1.tf")
+	testDataS2 := utils.ReadTestDataFile(t, "testdata/managed_object_storage_network_change_s2.tf")
+
+	objsto := managedObjectStorageResource
+	prefix := fmt.Sprintf("tf-acc-test-objsto-swap-%s-", acctest.RandString(4))
+	octet := acctest.RandIntRange(10, 250)
+	variables := map[string]config.Variable{
+		"prefix": config.StringVariable(prefix),
+		"cidr_a": config.StringVariable(fmt.Sprintf("172.%d.1.0/24", octet)),
+		"cidr_b": config.StringVariable(fmt.Sprintf("172.%d.2.0/24", octet)),
+	}
+
+	resource.ParallelTest(t, resource.TestCase{
+		PreCheck: func() { upcloud.TestAccPreCheck(t) },
+		Steps: []resource.TestStep{
+			{
+				ProtoV6ProviderFactories: upcloud.TestAccProviderFactories,
+				Config:                   testDataS1,
+				ConfigVariables:          variables,
+			},
+			{
+				ProtoV6ProviderFactories: upcloud.TestAccProviderFactories,
+				Config:                   testDataS2,
+				ConfigVariables:          variables,
+				ConfigPlanChecks: resource.ConfigPlanChecks{
+					PreApply: []plancheck.PlanCheck{
+						plancheck.ExpectResourceAction(objsto, plancheck.ResourceActionUpdate),
+					},
+				},
+				Check: resource.ComposeTestCheckFunc(
+					resource.TestCheckResourceAttr(objsto, "network.#", "2"),
+				),
+			},
+			{
+				ProtoV6ProviderFactories: upcloud.TestAccProviderFactories,
+				Config:                   testDataS2,
+				ConfigVariables:          variables,
+				ConfigPlanChecks: resource.ConfigPlanChecks{
+					PreApply: []plancheck.PlanCheck{
+						plancheck.ExpectResourceAction(objsto, plancheck.ResourceActionNoop),
 					},
 				},
 			},